ensure rclip can preview RAW images
